requirements.txt
# AI Survival - Local LLM Assistant
# Core dependencies

# LLM inference: uses llama-server.exe / llama-cli.exe directly
# llama-cpp-python not required — inference via subprocess + httpx

# Model downloading
huggingface-hub>=0.20.0

# Terminal UI
rich>=13.7.0

# Configuration
pyyaml>=6.0.1

# RAG / Knowledge base (FAISS + ask.py)
httpx>=0.27.0
faiss-cpu>=1.8.0
numpy>=1.26.0

# Utilities
tqdm>=4.66.0
